Modified to handle a sticky post
I am using this on the Linux Blade discord server, and one of the problems I was running into is that the Linux Corner forum has a sticky post as the first post. I modified the script to store the list of current IDs that it sees instead of just the last one.